Trim internal frames from AssertionFailedError
When using `JUnit.start` and creating a failing test, users
will be confronted with a large stacktrace with mostly irrelevant
information. Even after #5158 is merged, the stacktrace will contain
several internal frames:

```
org.opentest4j.AssertionFailedError: expected: <11> but was: <12>
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.AssertionFailureBuilder.build(AssertionFailureBuilder.java:158)
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.AssertionFailureBuilder.buildAndThrow(AssertionFailureBuilder.java:139)
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.AssertEquals.failNotEqual(AssertEquals.java:201)
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.AssertEquals.assertEquals(AssertEquals.java:152)
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.AssertEquals.assertEquals(AssertEquals.java:147)
	at org.junit.jupiter.api@6.1.0-SNAPSHOT/org.junit.jupiter.api.Assertions.assertEquals(Assertions.java:558)
	at com.examp.project/com.example.project.HelloTest.stringLength(HelloTest.java:14)
```

By pruning these internal frames, the stacktrace can be reduced to a
much more readable:

```
org.opentest4j.AssertionFailedError: expected: <11> but was: <12>
	at com.examp.project/com.example.project.HelloTest.stringLength(HelloTest.java:14)
```

Comparable behaviour can be found in AssertJ[1] and IDEA which folds
internal frames in the console using `<6 internal line>`.

The pruning functionality is intentionally added to the
`AssertionFailureBuilder` rather than the `ExceptionUtils` to enable
other users of the builder to also prune the internal frames from their
own assertions.

1. https://github.com/assertj/assertj/blob/79bdebf1817692e5e0ff5ee3ab097dcd104d47ae/assertj-core/src/main/java/org/assertj/core/util/Throwables.java#L117-L148